[TIR][Hexagon] Use the "target" value in T.func_attr for VTCM limit
For the VerifyVTCMLimit, read directly from the function attribute, if
the function has already been annotated with the target.
