+++ src/auto_scheduler/feature.cc
@@ -1408,9 +1408,7 @@ void GetPerStoreFeaturesWorkerFunc(const SearchTask& task, const State& state, i
     }
     if (IsHexagonTask(task)) {
       Target target = task->target;
-      const auto vtcm_capacity = target->GetAttr<Integer>("vtcm-capacity").value().IntValue();
-      const auto& optimize =
-          tir::transform::Sequential({tir::transform::VerifyVTCMLimit(vtcm_capacity)});
+      const auto& optimize = tir::transform::Sequential({tir::transform::VerifyVTCMLimit(target)});
       optimize(mod);
     }
     const auto& optimize =
